First timers and the imformation offered at easyy access?
Show more colors, symbols, and signs! Come on MAN THIS IS 2022! AI or UI understanding of multilingual languages means we no longer have to use txt. Or JavaScript or whatever else ABC 123 simple typing we use??? Use the GIF, stickers, and Emoji +++++ to talk to the machinery or shall I say human digi healthcode imprints????